Create Discord-Token-GRAPER
"This program is intended solely for ethical and educational purposes. It should only be used in environments where you have explicit permission to test or analyze systems. Unauthorized use of this software to access or extract sensitive information without consent is illegal, unethical, and a violation of privacy. Always adhere to legal guidelines and respect the rights and privacy of others."